Convert IROBOT to IPILOT (for SMartBot in QLearning)

+++ Umshini/Reinforcement/IState.cs
@@ -15,12 +15,12 @@ namespace Reinforcement
         /// <param name="agent"></param>
         /// <param name="opponent"></param>
         /// <returns>Integer representing the current state of the environment</returns>
-        double ConvertNumber(IROBOT agent, IROBOT opponent);
+        int ConvertNumber(IROBOT agent, IROBOT opponent);
 
         /// <summary>
         /// Gives the highest integer that can be obtained with this type of method
         /// </summary>
         /// <returns>Highest integer that can be obtained with this type of method</returns>
-        double GetMaxValue();
+        int GetMaxValue();
     }
 }
